benchmark/python/requirements.txt
python-docx>=0.8.11
docxtpl>=0.16.7
pytest>=7.0.0
pytest-benchmark>=4.0.0
memory-profiler>=0.60.0
psutil>=5.9.0
matplotlib>=3.5.0
pandas>=1.3.0
seaborn>=0.11.0
tabulate>=0.9.0 